fixed route stage

+++ src/services/stages.services.ts
@@ -1,19 +1,24 @@
-import { LEVEL_DATA } from './../utils/tools';
 import { Level } from 'types/types';
 import AppError from '../utils/appError';
 import { Projects, Stages, prisma } from '../utils/prisma.server';
 import { calculateAndUpdateDataByLevel } from '../utils/tools';
 import LevelsServices from './levels.services';
+import { existsSync } from 'fs';
+import PathLevelServices from './path_levels.services';
 
 class StageServices {
-  static async findMany() {
+  static async findMany(projectId: Projects['id']) {
     const findStages = await prisma.stages.findMany({
-      select: {
-        id: true,
-        name: true,
-      },
+      where: { projectId },
+      select: { id: true, name: true },
     });
     return findStages;
+  }
+  static async findShort(id: Stages['id']) {
+    if (!id) throw new AppError('Oops!, ID invalido', 400);
+    const findStage = await prisma.stages.findUnique({ where: { id } });
+    if (!findStage) throw new AppError('Oops!, ID invalido', 400);
+    return findStage;
   }
   static async find(id: Stages['id']) {
     if (!id) throw new AppError('Oops!, ID invalido', 400);
@@ -63,9 +68,10 @@ class StageServices {
         price: 0,
         level: 0,
         rootLevel: 0,
-        unique: false,
         stagesId: 0,
         userId: null,
+        isArea: false,
+        isProject: false,
         details: {
           UNRESOLVED: 0,
           PROCESS: 0,
@@ -83,25 +89,38 @@ class StageServices {
   }
 
   static async create({ name, projectId }: Stages) {
+    const path = await PathLevelServices.pathProject(projectId, 'UPLOADS');
+    if (!existsSync(path)) throw new AppError('Ops!,carpeta no existe', 404);
     const createStage = await prisma.stages.create({
       data: { name, projectId },
+      include: {
+        project: { select: { name: true } },
+      },
     });
     return createStage;
   }
 
   static async update(id: Stages['id'], { name }: Stages) {
     if (!id) throw new AppError('Oops!, ID invalido', 400);
+    const path = await PathLevelServices.pathStage(id, 'UPLOADS');
+    if (!existsSync(path)) throw new AppError('Ops!,carpeta no existe', 404);
     const updateStage = await prisma.stages.update({
       where: { id },
       data: { name },
+      include: {
+        project: { select: { name: true } },
+      },
     });
     return updateStage;
   }
 
   static async delete(id: Stages['id']) {
     if (!id) throw new AppError('Oops!, ID invalido', 400);
+    const path = await PathLevelServices.pathStage(id, 'UPLOADS');
+    if (!existsSync(path)) throw new AppError('Ops!,carpeta no existe', 404);
     const deleteStage = await prisma.stages.delete({
       where: { id },
+      include: { project: { select: { name: true } } },
     });
     return deleteStage;
   }
